+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 3.0.16 June 14, 2018
+Made a change in `dark/proteins.py`, to make the `minProteinFraction` work on
+a per sample basis, not per pathogen.
+
 ## 3.0.15 June 12, 2018
 
 Fixed another bug (unreference variable) in `graphics.py` that crept in in